% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp-package.R
\docType{package}
\name{adcp-package}
\alias{adcp-package}
\alias{adcp}
\title{Functions to Import and Analyse ADCP Files}
\description{
ADCP data from Sontek devices can be exported as Matlab (Mat)
files. In order to work with such files in R, this package allows
generic import functionality for such files. In addition, a plot
function is added that allows visualisation of relevant data.
Note that this package is currently under development and it is very
likely that it contains a rish variety of bugs and missing functionalities.
}
\details{
\tabular{ll}{ Package: \tab adcp\cr Type: \tab Package\cr Version: \tab
0.1.0\cr Date: \tab 2020-05-20\cr License: \tab GPL-3\cr }
}
\author{
Michael Dietze
}
\keyword{package}
