% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_add_sensor_frequency.R
\name{adcp_add_sensor_frequency}
\alias{adcp_add_sensor_frequency}
\title{Add column of ADCP frequency}
\usage{
adcp_add_sensor_frequency(dat)
}
\arguments{
\item{dat}{Data frame including column \code{sensor_model} with entries
"Sentinel_V20", "Sentinel_V50", "Sentinel_V100", and "Workhorse Sentinel
600kHz".}
}
\value{
Returns dat with additional column adcp_frequency_khz.
}
\description{
Add column of ADCP frequency
}
