% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_test_human_in_loop.R
\name{adcp_apply_human_in_loop}
\alias{adcp_apply_human_in_loop}
\title{Apply human in the loop flags and comments}
\usage{
adcp_apply_human_in_loop(
  dat,
  human_in_loop_table = NULL,
  qc_tests = c("tidal_bin_height", "grossrange")
)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in wide format.}

\item{human_in_loop_table}{One row of data table with information required to
assign the human in the loop flags. Must include the following columns:
station, depl_id, variable, timestamp_utc_min,    timestamp_utc_max,
qc_test_column (qc flag column that is being upgraded), qc_flag_value
(existing flag value for qc_test_column), human_in_loop_flag_value, and
human_in_loop_comment (will be included in the dataset).}

\item{qc_tests}{Character vector of quality control tests that have been
applied to \code{dat}. Passed to \code{adcp_pivot_flags_longer()}. Default
is: \code{qc_tests = c("bin_height", "grossrange")}.}
}
\value{
Returns \code{dat} with a \code{human_in_loop_flag} column for each
  variable and a \code{human_in_loop_comment} column.
}
\description{
Apply human in the loop flags and comments
}
