% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_assign_altitude.R
\name{adcp_assign_altitude}
\alias{adcp_assign_altitude}
\title{Assign altitude (height above sea floor) to each bin}
\usage{
adcp_assign_altitude(
  dat,
  metadata = NULL,
  inst_alt = NULL,
  bin_size = NULL,
  first_bin_range = NULL
)
}
\arguments{
\item{dat}{Output from \code{read_adcp_txt()}.}

\item{metadata}{Dataframe with metadata information for the deployment in
\code{dat} (e.g., a row from the NSDFA tracking sheet). Must include
columns \code{Inst_Altitude}, \code{Bin_Size}, and \code{First_Bin_Range}.
If default value \code{metadata = NULL} is used, the remaining arguments
must be provided.}

\item{inst_alt}{Height of the instrument above the sea floor (in metres). Not
required if \code{metadata} argument is specified.}

\item{bin_size}{Size of each bin (in metres). Not required if \code{metadata}
argument is specified.}

\item{first_bin_range}{Size of the first bin (in metres). Not required if
\code{metadata} argument is specified.}
}
\value{
Returns \code{dat}, with bin columns re-named with corresponding
  altitude (in metres).
}
\description{
Assign altitude (height above sea floor) to each bin
}
