% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_calculate_bin_depth.R
\name{adcp_calculate_bin_depth}
\alias{adcp_calculate_bin_depth}
\title{Calculate the bin depth below the surface}
\usage{
adcp_calculate_bin_depth(dat, metadata = NULL, inst_alt = NULL)
}
\arguments{
\item{dat}{Dataframe of ACDP data in long format, as returned by
\code{adcp_pivot_longer()}.}

\item{metadata}{Dataframe with metadata information for the deployment in
\code{dat} (e.g., a row from the NSDFA tracking sheet). Must include column
Inst_Altitude. Option to use default value \code{metadata = NULL} and
provide the required value in the \code{inst_alt} argument.}

\item{inst_alt}{Height of the instrument above the sea floor (in metres). Not
used if \code{metadata} argument is specified.}
}
\description{
Calculate the bin depth below the surface
}
\details{
Bin depth below the surface is calculated as:

  bin_depth_below_surface_m = sensor_depth_below_surface_m + inst_alt -
  bin_height_above_sea_floor_m

  A warning is printed if any bin_depth_below_surface_m are negative.
}
