% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_calc_vector_average.R
\name{adcp_calculate_vector_average}
\alias{adcp_calculate_vector_average}
\title{Generate summary table of flags}
\usage{
adcp_calculate_vector_average(
  dat,
  ...,
  speed_col = sea_water_speed_cm_s,
  direction_degree_col = sea_water_to_direction_degree,
  return_components = FALSE
)
}
\arguments{
\item{dat}{Data frame with columns for speed and direction and optional
grouping arguments.}

\item{...}{Optional argument. Column names (not quoted) from \code{dat} to
use as grouping variables.}

\item{speed_col}{The column in \code{dat} that holds the speed data (NOT
QUOTED).}

\item{direction_degree_col}{The column in \code{dat} that holds the direction
data (NOT QUOTED). Must be in units of degrees from vertical.}

\item{return_components}{Logical argument. If \code{TRUE} the x and y vector
components.}
}
\value{
Returns \code{dat} with additional columns.
}
\description{
Calculate the vector average for current speed and direction.
}
\details{
Step 1: calculate components of each direction vector weighted by speed.

\deqn{vx = speed * sin(direction_{degree} * pi / 180)}

\deqn{vy = speed * cos(direction_{degree} * pi / 180)}

Step 2: Calculate average of each component.

\deqn{vx_{mean} = mean(vx)}

\deqn{vy_{mean} = mean(vy)}

Step 3: Calculate new vector based on the mean components.

\deqn{direction_{mean} = atan2(vx_{mean}, vy_{mean}) * 180/pi}

\deqn{direction_{mean} = (360 + direction_{mean}) \%\% 360}

\deqn{speed_{mean} = sqrt(vy_{mean}^2 + vx_{mean}^2)}
}
