% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_compile_deployment_info.R
\name{adcp_compile_deployment_info}
\alias{adcp_compile_deployment_info}
\title{Export ADCP Data Deployment Information Dataset}
\usage{
adcp_compile_deployment_info(path_nsdfa = NULL, sheet = NULL)
}
\arguments{
\item{path_nsdfa}{Full file path for the nsdfa tracking sheet, including file
name and extension. If \code{NULL}, will default to the 2023-11-27 version
on the CMAR R drive.}

\item{sheet}{Name of the sheet to read in as metadata. Defaults is
"CurrMetaData".}
}
\value{
Exports data frame of wave deployment information.
}
\description{
Imports NSDFA Tracking sheet and ADCP TRACKING and exports the
  metadata for the Nova Scotia Open Data Portal.
}
\details{
TO DO: re-write this with new tracking sheet
}
