% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adcp_convert_bin_height_to_ordered_factor}
\alias{adcp_convert_bin_height_to_ordered_factor}
\title{Convert bin_height_above_sea_floor_m column to ordered factor}
\usage{
adcp_convert_bin_height_to_ordered_factor(dat)
}
\arguments{
\item{dat}{Data frame that includes the column
\code{bin_height_above_sea_floor_m}.}
}
\value{
Returns data with \code{bin_height_above_sea_floor_m} as an ordered
  factor.
}
\description{
Depths are assigned levels in descending order so that figures will have
shallowest bins at the top and deepest bins at the bottom of the figure.
}
