% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_bin_obs.R
\name{adcp_count_obs}
\alias{adcp_count_obs}
\title{Count number and proportion of observations in equal intervals}
\usage{
adcp_count_obs(dat_speed, n_ints = 12, n_digits = 2, label_sep = "\\n")
}
\arguments{
\item{dat_speed}{Data observations to assign intervals. The interval size is determined by the
range of these observations and \code{n_ints}.}

\item{n_ints}{Number of intervals to divide observations into. The interval size is
determined by the range of these observations and \code{n_ints}.}

\item{n_digits}{Number of digits to show in interval limits.}

\item{label_sep}{Separator for the interval labels ("lower to upper"). Default is
a new line to save room on plot axis.}
}
\value{
Returns a dataframe of upper and lower interval limits (lower inclusive),
  and frequency and proportion of observations in each interval.
}
\description{
Count number and proportion of observations in equal intervals
}
