% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp-package.R
\docType{data}
\name{adcp_ex}
\alias{adcp_ex}
\alias{x}
\title{Example ADCP data set}
\format{
The format is: List of 20
                         $ :List of 23
                         ..$ xy        : num [1:2] 539985 7758785
                         ..$ track     : num [1:2] -0.048 0.025
                         ..$ depth     : num [1:22] 0.13 0.15 0.17 ...
                         ..$ depth_plot: num [1:23] 0.12 0.14 0.16 ...
                         ..$ angle     : num 340
                         ..$ vel_l     : num [1:22] 0.441 0.452 0.471 ...
                         ..$ vel_x     : num [1:22] -0.0044 0.0239 0.16...
                         ..$ vel_v     : num [1:22] -0.012 -0.011 0.026 ...
                         ..$ vel_s     : num [1:22] 0.442 0.453 0.5 0.483 ...
                         ..$ vel_e     : num [1:22] -0.153 -0.13 -0.005 ...
                         ..$ vel_n     : num [1:22] 0.414 0.434 0.499 ...
                         ..$ vel_u     : num [1:22] -0.012 -0.011 0.026  ...
                         ..$ vel_d     : num [1:22] -0.044 -0.021 0.063 ...
                         ..$ snr_e     : num [1:22] 21.9 44.1 46.3 44.6 ...
                         ..$ snr_n     : num [1:22] 22.8 39.4 41.3 47.4 ...
                         ..$ snr_u     : num [1:22] 35.5 45.1 47.3 43.6 ...
                         ..$ snr_d     : num [1:22] 32.1 42.5 42.5 44.2 ...
                         ..$ snr_mean  : num [1:22] 28.1 42.8 44.3 44.9 ...
                         ..$ cor_e     : num [1:22] 54 60 62 37 23 30 40 ...
                         ..$ cor_n     : num [1:22] 31 41 18 19 28 1 17 ...
                         ..$ cor_u     : num [1:22] 69 63 60 60 51 39 37 ...
                         ..$ cor_d     : num [1:22] 31 50 43 61 56 64 46 ...
                         ..$ cor_mean  : num [1:22] 46.2 53.5 45.8 44.2 ...
}
\usage{
x
}
\description{
The dataset contains 20 measurements of a Sontek sensor from a 0.6 m
deep river. The file has been imported to R using the function
\code{adcp::read_adcp}.
}
\examples{

## load example data set
data(adcp_ex)

## plot stream-wise velocity
plot_adcp(x = x, what = "vel_l")

}
\keyword{datasets}
