% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adcp_extract_deployment_info}
\alias{adcp_extract_deployment_info}
\title{Extract deployment date and station name from file path}
\usage{
adcp_extract_deployment_info(file_path)
}
\arguments{
\item{file_path}{Path to the file, include file name and extension (.csv or
.txt). File name must include the deployment date and the station name, in
the format YYYY-MM-DD_Station Name.ext (e.g., 2007-12-18_Spectacle
Island.txt)}
}
\value{
Returns a tibble with three columns: \code{DEPLOYMENT},
  \code{Depl_Date}, and \code{Station_Name}.
}
\description{
Extract deployment date and station name from file path
}
