% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adcp_extract_deployment_info2}
\alias{adcp_extract_deployment_info2}
\title{Extract deployment date and station name from compiled rds file}
\usage{
adcp_extract_deployment_info2(file_path)
}
\arguments{
\item{file_path}{Path to the file, include file name and extension (.rds).
File name must include the deployment date, the station name, and the
deployment id separated by "_", e.g., "2020-10-22_Cornwallis_SW_AN002".}
}
\value{
Returns a data frame with three columns: \code{depl_date},
  \code{station}, and \code{deployment_id}.
}
\description{
Extract deployment date and station name from compiled rds file
}
