% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_import_data.R
\name{adcp_import_data}
\alias{adcp_import_data}
\title{Import current data from rds files.}
\usage{
adcp_import_data(path_input = NULL, county = "all", add_county_col = TRUE)
}
\arguments{
\item{path_input}{Path to the *.rds files to be imported. Default is the
ADCP/Open Data/County Datasets folder on the CMAR Operations shared drive
(user must be connected to the Perennia VPN).}

\item{county}{Vector of character string(s) indicating which county or
counties for which to import data. The filter is applied to the file path, so
the county name MUST be part of the file path (e.g., the name of the folder).
Defaults to all counties.}

\item{add_county_col}{Logical argument indicating whether to include a
"county" column in the output. If \code{TRUE}, the imported data must have
"waterbody" and "station" columns.}
}
\description{
Import current data from rds files.
}
