% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_label_speed.R
\name{adcp_label_speed}
\alias{adcp_label_speed}
\title{Assign intervals to speed data}
\usage{
adcp_label_speed(
  dat,
  column = sea_water_speed_cm_s,
  n_ints = 12,
  upper_value = NULL,
  label_sep = " "
)
}
\arguments{
\item{dat}{Data frame including the column with observations that will be
assigned to intervals. The interval size is determined by the range of
these observations and \code{n_ints}.}

\item{column}{Column in \code{dat} that will be assigned to intervals for
frequency table (NOT QUOTED).}

\item{n_ints}{Number of intervals to divide observations into. The interval
size is determined by the range of these observations and \code{n_ints}.
Alternatively, a vector of two or more unique break points. Passed to the
\code{breaks} argument of \code{cut()}.}

\item{upper_value}{Value used to generate the upper limit of the final
interval. Default is the maximum values of \code{ column }.}

\item{label_sep}{Separator for the interval labels ("lower to upper").
Default is a new line to save room on plot axis.}
}
\value{
Returns \code{dat} with an additional column \code{column_labels},
  the interval labels as an ordered factor.
}
\description{
Assigns intervals to \code{column} using the \code{cut()} function.
\code{column} will typically be a \code{sea_water_speed} column. To label
direction data, use \code{adcp_label_direction()}.
}
\details{
Intervals start and 0 and end at the smallest value that includes all
observations and results in \code{n_ints} equal intervals. Some small and/or
large intervals may have zero observations.
}
