% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_pivot_bin_height.R
\name{adcp_pivot_bin_height}
\alias{adcp_pivot_bin_height}
\title{Pivot ADCP data from bin height as column names to bin height as observations}
\usage{
adcp_pivot_bin_height(dat_wide, rm_NA = TRUE)
}
\arguments{
\item{dat_wide}{Data frame of ADCP data, as returned from
\code{adcp_read_txt()} or \code{adcp_assign_alt()}.}

\item{rm_NA}{Logical argument. If \code{rm_NA = TRUE}, rows where
\code{sea_water_speed_m_s} is \code{NA} OR
\code{sea_water_to_direction_degree} is \code{NA} will be removed.}
}
\value{
Returns data in a long format.

  If \code{dat_wide} is from \code{adcp_read_txt()} (i.e., bin altitude has
  not been assigned), then the following columns are returned: timestamp_utc,
  sensor_depth_below_surface_m (SensorDepth), bin_id (the default column
  names when imported using read_adcp_txt; starts at V8), sea_water_speed_m_s
  (WaterSpeed), and sea_water_to_direction_degree (WaterDirection).

  If \code{dat_wide} is from \code{adcp_assign_altitude()}, then the
  following columns are returned: timestamp_utc, sensor_depth_below_surface_m
  (SensorDepth), bin_depth_below_surface_m, bin_height_above_sea_floor_m,
  sea_water_speed_m_s (WaterSpeed), and sea_water_to_direction_degree
  (WaterDirection).
}
\description{
Pivot ADCP data from bin height as column names to bin height as observations
}
