% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_pivot_flags_wider.R
\name{adcp_pivot_flags_wider}
\alias{adcp_pivot_flags_wider}
\title{Pivot flagged current data wider by variable}
\usage{
adcp_pivot_flags_wider(dat, qc_tests = NULL)
}
\arguments{
\item{dat}{Data frame of flagged current data in long format.}

\item{qc_tests}{Quality control tests included in \code{dat_wide}. Default is
\code{qc_tests = c("tidal_bin_height", "grossrange", "qc")}. Will also work
for "rolling_sd" and "spike". If \code{dat} only includes the max flag, use
\code{qc_tests = "qc"}.}
}
\value{
Returns \code{dat}, with variables and flags pivoted to a wide
  format, i.e., one qc column for each variable and test EXCEPT
  tidal_bin_height_flag, which applies to all variables.
}
\description{
Pivot flagged current data wider by variable
}
