% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_pivot_flags_longer.R
\name{adcp_pivot_single_test_longer}
\alias{adcp_pivot_single_test_longer}
\title{Complete pivot_longer of flagged current data}
\usage{
adcp_pivot_single_test_longer(dat_wide, qc_test)
}
\arguments{
\item{dat_wide}{Data frame of flagged current data with the variables
in a long format and flags in wide format.}

\item{qc_test}{Flag columns to pivot.}
}
\value{
Returns \code{dat_wide} with the qc_test flag columns pivoted to a
  long format.
}
\description{
Complete pivot_longer of flagged current data
}
