% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_pivot_vars_longer.R
\name{adcp_pivot_vars_longer}
\alias{adcp_pivot_vars_longer}
\title{Pivot current data}
\usage{
adcp_pivot_vars_longer(dat_wide, vars = NULL)
}
\arguments{
\item{dat_wide}{Data frame of ADCP current data.}

\item{vars}{Vector of character strings indicating which columns to pivot.
Default is all variables.}
}
\value{
Returns data in a long format.
}
\description{
Pivot current data
}
