% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_current_speed_time.R
\name{adcp_plot_current_speed_time}
\alias{adcp_plot_current_speed_time}
\title{Plot histogram of speed observations}
\usage{
adcp_plot_current_speed_time(dat)
}
\arguments{
\item{dat}{Data frame that includes columns \code{station},
\code{bin_height_above_sea_floor_m}, \code{sea_water_speed_cm_s} and
\code{timestamp_utc}.}
}
\value{
Returns a ggplot object.
}
\description{
Plot histogram of speed observations
}
