% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_depth_flags.R
\name{adcp_plot_depth}
\alias{adcp_plot_depth}
\title{Plot sensor_depth_below_surface_m}
\usage{
adcp_plot_depth(dat, title = NULL, date_format = "\%Y-\%b-\%d", geom = "point")
}
\arguments{
\item{dat}{Dataframe of ACDP data in long format, as returned by
\code{adcp_format_opendata()}.}

\item{title}{Optional title for the figure.}

\item{date_format}{Format for the date labels. Default is "\%Y-\%b-\%d"}

\item{geom}{Geom to plot. Options are \code{"point"} or \code{"line"}.}
}
\value{
ggplot object. Figure shows sensor_depth_below_surface_m over time.
}
\description{
Plot sensor_depth_below_surface_m
}
