% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_depth_flags.R
\name{adcp_plot_depth_flags}
\alias{adcp_plot_depth_flags}
\title{Plot SENSOR_DEPTH colour-coded by FLAG}
\usage{
adcp_plot_depth_flags(dat, title = NULL)
}
\arguments{
\item{dat}{same same}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot SENSOR_DEPTH colour-coded by FLAG
}
