% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_var_time.R
\name{adcp_plot_var_time}
\alias{adcp_plot_var_time}
\title{Plot histogram of speed observations}
\usage{
adcp_plot_var_time(
  dat,
  plot_col,
  y_axis_label = NULL,
  geom = "line",
  date_format = "\%Y-\%b-\%d",
  pal = "#000000"
)
}
\arguments{
\item{dat}{Data frame that includes columns \code{timestamp_utc} (POSIXct),
and \code{plot_col} (numeric).}

\item{plot_col}{Column to plot (UNQUOTED).}

\item{y_axis_label}{Character string for the y-axis title. Could hard code
some options.}

\item{geom}{Character value indicating which geom to use. Options are "point"
or "line."}

\item{date_format}{Character string indicating the format for the date
labels.}

\item{pal}{Single colour value (hex code or number).}
}
\value{
Returns a ggplot object.
}
\description{
Plot histogram of speed observations
}
