% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_read_nsdfa_metadata.R
\name{adcp_read_nsdfa_metadata}
\alias{adcp_read_nsdfa_metadata}
\title{Import NSDFA tracking sheet and extra deployment metadata}
\usage{
adcp_read_nsdfa_metadata(
  path,
  sheet = NULL,
  station = NULL,
  deployment_date = NULL
)
}
\arguments{
\item{path}{Path to the NSDFA tracking sheet (include file name and
extension).}

\item{sheet}{Sheet to read in. Defaults to the first sheet.}

\item{station}{Station for which to return metadata.}

\item{deployment_date}{Date of deployment for which to return metadata.}
}
\value{
Returns data frame of NSDFA tracking sheet ADCP metadata. Option to
  filter for a single deployment.
}
\description{
Import NSDFA tracking sheet and extra deployment metadata
}
\details{
Reads in the NSDFA tracking sheet and corrects known errors (e.g.,
  standardizes station and waterbody spellings, fixes deployment dates, and
  etc.).
}
