% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_setup_folders.R
\name{adcp_set_up_folders}
\alias{adcp_set_up_folders}
\title{Set up folder structure for processing ADCP data}
\usage{
adcp_set_up_folders(path = NULL, folder = NULL)
}
\arguments{
\item{path}{Path to where the folder will be created. Default is the "Side
Lobe Trimmed" folder on the Perennia shared drive.}

\item{folder}{Name of the top-level folder to be created. Default is
todays_date_Process.}
}
\value{
Returns folder structure that works well with CMAR ADCP processing
  templates.
}
\description{
Set up folder structure for processing ADCP data
}
