% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_test_grossrange.R
\name{adcp_test_grossrange}
\alias{adcp_test_grossrange}
\title{Apply grossrange test to current variables}
\usage{
adcp_test_grossrange(
  dat,
  current_grossrange_table = NULL,
  county = NULL,
  vars = NULL
)
}
\arguments{
\item{dat}{Data frame of current variables in wide format.}

\item{current_grossrange_table}{Data frame with at least 5 columns:
\code{variable}: entries must match the names of the variables being tested
in \code{dat}; \code{gr_min}: minimum acceptable value; \code{gr_max}:
maximum accepted value ; \code{user_min}: minimum reasonable value;
\code{user_max}: maximum reasonable value.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Used to filter the default \code{current_grossrange_table}.}

\item{vars}{Vector of character strings indicating which columns to pivot.
Default is all variables.}
}
\value{
Returns \code{dat} with an additional grossrange_flag column for each
  wave variable.
}
\description{
Apply grossrange test to current variables
}
