% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_trim_NA.R
\name{adcp_trim_NA}
\alias{adcp_trim_NA}
\title{Filter out ensembles with no speed or direction data}
\usage{
adcp_trim_NA(dat)
}
\arguments{
\item{dat}{Output from \code{read_adcp_txt()}.}
}
\value{
Returns \code{dat} filtered to exclude ensembles with no water speed
  or direction data (typically as the sensor is deployed and retrieved).
}
\description{
Filter out ensembles with no speed or direction data
}
