% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_trim_NA.R
\name{adcp_trim_NA_bins}
\alias{adcp_trim_NA_bins}
\title{Filter out bins with no depth, speed or direction data}
\usage{
adcp_trim_NA_bins(dat)
}
\arguments{
\item{dat}{Output from \code{read_adcp_txt()}.}
}
\value{
Returns \code{dat} filtered to exclude bins (columns) with no water
  speed or direction data (typically NA because of side lobe interference).
}
\description{
Filter out bins with no depth, speed or direction data
}
