% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_threshold_tables.R
\docType{data}
\name{current_thresholds}
\alias{current_thresholds}
\title{Threshold tables}
\format{
Default thresholds for adding flags to current data.
}
\usage{
current_thresholds
}
\description{
Threshold tables
}
\keyword{datasets}
