% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_speed_colour_pal}
\alias{get_speed_colour_pal}
\title{Generate colour palette for speed values}
\usage{
get_speed_colour_pal(n_colours)
}
\arguments{
\item{n_colours}{Integer indicating how many colours to generate.}
}
\value{
Character vector of \code{n_levels} hex codes from \code{viridis}
  option "F".
}
\description{
Generate colour palette for speed values
}
