% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_adcp.R
\name{interpolate_adcp}
\alias{interpolate_adcp}
\title{Interpolate (rasterise) ADCP measurements to equal intervals}
\usage{
interpolate_adcp(x, n, extend = TRUE)
}
\arguments{
\item{x}{\code{list}, ADCP data set to be interpolated}

\item{n}{\code{Numerical} value, number of output depth values. If
omitted, it is generated from the largest number of cells of the input
data set.}

\item{extend}{\code{Logical} value, option to provide values also
outside the original data range of individual measurements. Default
is \code{TRUE} (extension of interpolation outside original data range).}
}
\value{
Interpolated (rasterised) ADCP data set
}
\description{
This function interpolates cell-based ADCP data to equal intervals,
constant across all measurements.
}
\examples{

## load example data set
data(adcp_ex)

## interpolate and plot data set
x_int <- interpolate_adcp(x = x, n = 10)
plot_adcp(x = x_int)

## interpolate and plot data set without extrapolation
x_int_2 <- interpolate_adcp(x = x, n = 10, extend = FALSE)
plot_adcp(x = x_int_2)

}
\author{
Michael Dietze
}
\keyword{ADCP}
