% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_adcp.R
\name{list_adcp}
\alias{list_adcp}
\title{Output of adcp object elements}
\usage{
list_adcp()
}
\value{
List object with adcp object elements and short description.
}
\description{
Prints the elements of an imported ADCP data set, along with a short
description of the element names.
}
\examples{

list_adcp()

}
