% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_adcp.R
\name{matrix_adcp}
\alias{matrix_adcp}
\title{Convert matrix to ADCP list object element}
\usage{
matrix_adcp(x, adcp, what)
}
\arguments{
\item{x}{\code{Matrix} data set to be converted to ADCP object}

\item{adcp}{\code{ADCP} object to be changed. Object is a list as, for
example created by \code{rad_adcp}.}

\item{what}{\code{Character} value, name of the data set to convert. If
the same as one of the already existing ADCP object names, that will be
overwritten. Otherwise, the data will be appended as new element of the
ADCp object}
}
\value{
\code{ADCP} object.
}
\description{
Convert matrix to ADCP list object element
}
\examples{

x = 1

}
\author{
Michael Dietze
}
\keyword{ADCP}
