% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adcp.R
\name{plot_adcp}
\alias{plot_adcp}
\title{Plot ADCP data set}
\usage{
plot_adcp(x, what = "vel_l", ref = "ID", ...)
}
\arguments{
\item{x}{\code{list}, ADCP data set to be plotted. Should be generated
using \code{read_adcp} or \code{interpolate_adcp}.}

\item{what}{\code{Character} value, variable to be plotted. See
\code{list_adcp()} for available keywords. Default is \code{vel_l}
(velocity along stream).}

\item{ref}{\code{Character} value, x-axis reference. Can be one out of
\code{"track"} (GPS based track position), \code{xy} (GPS based
coordinates) and \code{ID} (measurement sequence ID). Default
is \code{"ID"}.}

\item{\dots}{Additional arguments passed to the plot function.}
}
\value{
Image plot of the data set.
}
\description{
Plots one of a series of selectable ADCP data set variables.
}
\details{
Available keywords are:

\tabular{ll}{
  \strong{keyword} \tab \strong{description}\cr
  xy \tab XY coordinates\cr
  track \tab Device track coordinates\cr
  depth \tab Measurement cell depth vector\cr
  depth_plot \tab Plotting depths (upper-lower boundaries)\cr
  angle \tab Device orientation angle\cr
  vel_l \tab Velocity along stream\cr
  vel_x \tab Velocity across stream\cr
  vel_v \tab Velocity vertical\cr
  vel_s \tab Velocity vector sum\cr
  vel_e \tab Velocity east component\cr
  vel_n \tab Velocity north component\cr
  vel_u \tab Velocity up component\cr
  vel_d \tab Velocity difference component\cr
  snr_e \tab SNR east component\cr
  snr_n \tab SNR north component\cr
  snr_u \tab SNR up component\cr
  snr_d \tab SNR difference component\cr
  snr_mean \tab SNR average\cr
  cor_e \tab Correlation east component\cr
  cor_n \tab Correlation north component\cr
  cor_u \tab Correlation up component\cr
  cor_d \tab Correlation difference component\cr
  cor_mean \tab Correlation average\cr
}

Data sets that have been subject to the function \code{difference_adcp}
will contain the following additional elements:

\tabular{ll}{
  \strong{keyword} \tab \strong{description}\cr
  vel_l_diff \tab Velocity along stream, deviation from average\cr
  vel_x_diff \tab Velocity across stream, deviation from average\cr
  vel_v_diff \tab Velocity vertical, deviation from average\cr
  vel_s_diff \tab Velocity vector sum, deviation from average\cr
  vel_e_diff \tab Velocity east component, deviation from average\cr
  vel_n_diff \tab Velocity north component, deviation from average\cr
  vel_u_diff \tab Velocity up component, deviation from average\cr
  vel_d_diff \tab Velocity difference component, deviation from average\cr
}
}
\examples{

## load example data
data(adcp_ex)

## plot data set
plot_adcp(x = x)

## plot data set with adjusted scale limits
plot_adcp(x = x,
          xlim = c(5, 10),
          ylim = c(0.5, 0.3),
          zlim = c(0.2, 0.5))

## plot data set with customised axes labels
plot_adcp(x = x,
          xlab = "Measurement ID",
          ylab = "Depth",
          zlab = "Velocity along stream (m/s)")

## plot data set with track as reference
plot_adcp(x = x,
          ref = "track")

## plot across-stream velocity
plot_adcp(x = x, what = "vel_x")

## create customised colour scale
col_user <- colorRampPalette(c("grey20", "grey50", "red",
                               "orange", "yellow", "white"))

## plot data set with customised colour scale
plot_adcp(x = x, col = col_user(200))

}
\author{
Michael Dietze
}
\keyword{ADCP}
