% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_assign_altitude.R
\name{adcp_assign_altitude}
\alias{adcp_assign_altitude}
\title{Assign altitude (height above sea floor) to each bin}
\usage{
adcp_assign_altitude(
  dat_wide,
  metadata = NULL,
  inst_alt = NULL,
  bin_size = NULL,
  first_bin_range = NULL
)
}
\arguments{
\item{dat_wide}{Data frame of ADCP data, as exported from
\code{adcp_read_txt()}.}

\item{metadata}{Data frame with metadata information for the deployment in
\code{dat_wide} (e.g., a row from the NSDFA tracking sheet). Must include
columns \code{Inst_Altitude}, \code{Bin_Size}, and \code{First_Bin_Range}.
Option to use default value \code{metadata = NULL} and provide the required
values in the remaining arguments.}

\item{inst_alt}{Height of the sensor above the sea floor (in metres). Not
used if \code{metadata} argument is specified.}

\item{bin_size}{Size of each bin (in metres). Not used if \code{metadata}
argument is specified.}

\item{first_bin_range}{Distance from the transducer face to the centre of the
first bin (in metres). Not used if \code{metadata} argument is specified.}
}
\value{
Returns \code{dat_wide}, with bin columns re-named with corresponding
  altitude (in metres).
}
\description{
Assign altitude (height above sea floor) to each bin
}
\details{
The height to the centre of the bins is calculated as:

  first bin altitude = inst altitude + first bin range

  second bin altitude = first bin altitude + bin size

  third bin altitude = second bin altitude + bin size

  And so on.
}
