% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_check_duplicate_timestamp.R
\name{adcp_check_duplicate_timestamp}
\alias{adcp_check_duplicate_timestamp}
\title{Check for duplicate TIMESTAMP values}
\usage{
adcp_check_duplicate_timestamp(dat)
}
\arguments{
\item{dat}{Dataframe, as exported from \code{adcp_read_txt()}}
}
\value{
Returns a warning if any duplicate values are detected.
}
\description{
Check for duplicate TIMESTAMP values
}
