% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adcp_check_new_folder}
\alias{adcp_check_new_folder}
\title{Check if there are files in the specified folder}
\usage{
adcp_check_new_folder(path, pattern = "csv")
}
\arguments{
\item{path}{Path to the folder to check.}

\item{pattern}{Character string indicating what file type to check for.}
}
\value{
Returns a Warning if there are files in the folder. Returns a
  message if not.
}
\description{
Check if there are files in the specified folder
}
