% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adcp_convert_flag_to_ordered_factor}
\alias{adcp_convert_flag_to_ordered_factor}
\title{Convert FLAG column to ordered factor}
\usage{
adcp_convert_flag_to_ordered_factor(dat)
}
\arguments{
\item{dat}{Dataframe of ACDP data in long format, as returned by
\code{adcp_flag_data()}.}
}
\value{
Returns dat, with the FLAG column as an ordered factor, with levels
"good" < "SENSOR_DEPTH_BELOW_SURFACE changed by > x m" < "manual flag".
}
\description{
Convert FLAG column to ordered factor
}
