% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_correct_timestamp.R
\name{adcp_correct_timestamp}
\alias{adcp_correct_timestamp}
\title{Convert timestamp to UTC from AST or DST}
\usage{
adcp_correct_timestamp(dat)
}
\arguments{
\item{dat}{Dataframe with at least one column \code{TIMESTAMP_NS}.}
}
\value{
Returns \code{dat} with \code{TIMESTAMP} in UTC.
}
\description{
Convert timestamp to UTC from AST or DST
}
