% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_export_deployment_info.R
\name{adcp_export_deployment_info}
\alias{adcp_export_deployment_info}
\title{Export Current Data Deployment Information Dataset}
\usage{
adcp_export_deployment_info(deployments, path_nsdfa, path_export)
}
\arguments{
\item{deployments}{Vector of deployment IDs to include in the dataset.}

\item{path_nsdfa}{Full file path for the nsdfa tracking sheet, including file
name and extension.}

\item{path_export}{File path to the folder where the Deployment Information
Dataset should be exported.}
}
\value{
Exports csv file named todays-date_Current_Data_Deployment_Info.csv.
}
\description{
Imports NSDFA Tracking sheet and Deployment ID tracker and
  exports the metadata for the Nova Scotia Open Data Portal. User must be
  connected to CMAR shared drive.
}
