% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_flag_data.R
\name{adcp_flag_data}
\alias{adcp_flag_data}
\title{Flag ensembles with suspect SENSOR_DEPTH recordings}
\usage{
adcp_flag_data(dat, depth_flag = 1)
}
\arguments{
\item{dat}{The usual (formatted)}

\item{depth_flag}{The change in \code{SENSOR_DEPTH} that will trigger a flag.}
}
\value{
Returns dat with two extra columns for inspection: \code{DIFF} =
  lead(SENSOR_DEPTH) - SENSOR_DEPTH and \code{FLAG}.
}
\description{
Flag ensembles with suspect SENSOR_DEPTH recordings
}
