% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_format_opendata.R
\name{adcp_format_opendata}
\alias{adcp_format_opendata}
\title{Convert ADCP data from raw format to Open Data Portal format}
\usage{
adcp_format_opendata(dat)
}
\arguments{
\item{dat}{Output from  \code{adcp_read_txt()} or \code{adcp_assign_alt()}.}
}
\value{
Fill this in
}
\description{
Convert ADCP data from raw format to Open Data Portal format
}
