% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_format_report_table.R
\name{adcp_format_report_table}
\alias{adcp_format_report_table}
\title{Formats deployment table for summary report}
\usage{
adcp_format_report_table(report_table, transpose = TRUE)
}
\arguments{
\item{report_table}{Deployment metadata from the NSDFA tracking sheet, as
returned from \code{adcp_write_report_table()}.}

\item{transpose}{Logical argument indicating whether to transpose
\code{report_table} before applying the format. Use the default \code{TRUE}
for the deployment table. Set to \code{FALSE} for the document history
table.}
}
\value{
Returns a flextable object that will render nicely in the Word
  report. A table with two columns: the first column (bold) is the column
  names of \code{report_table}; the second column is the corresponding
  entries in \code{report_table}.
}
\description{
Formats deployment table for summary report
}
