% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_label_direction.R
\name{adcp_label_direction}
\alias{adcp_label_direction}
\title{Cut wind direction into 8 bins}
\usage{
adcp_label_direction(dat, n_petals = 16)
}
\arguments{
\item{dat}{Data frame with column sea_water_to_direction_degree.}

\item{n_petals}{Number of bins to divide direction data into. Must be either
8 or 16.}
}
\value{
Returns \code{dat} with an additional column
  \code{sea_water_to_direction_degree_labels}, the direction labels as an
  ordered factor.
}
\description{
Adapted from \code{cutVecWinddir()} from \code{OpenAir}.
}
\details{
Likely more efficient in the original vectorized form. Modified to match
\code{adcp_label_speed()}. Could re-visit both.

If \code{n_petals} is 8: Assigns direction data in bins of 45 degrees. -22.5
to 22.5 degree is North, 22.5 to 67.5 is NE, etc.

If \code{n_petals} is 16 (the default): Assigns direction data in bins of
22.5 degrees. -11.25 to 11.25 degree is North, 11.25 to 33.75 is NNE, etc.
}
