% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_label_direction.R
\name{adcp_label_direction}
\alias{adcp_label_direction}
\title{Cut wind direction into 8 bins}
\usage{
adcp_label_direction(dat)
}
\arguments{
\item{dat}{Data frame with column sea_water_to_direction_degree.}
}
\value{
Returns \code{dat} with an additional column
  \code{sea_water_to_direction_degree_labels}, the direction labels as an
  ordered factor.
}
\description{
Adapted from \code{cutVecWinddir()} from \code{OpenAir}.
}
\details{
Likely more efficient in the original vectorized form. Modified to match
\code{adcp_label_speed()}. Could re-visit both.

Assigns direction data in bins of 45 degree C. -22.5 to 22.5 degree is North,
22.5 to 67.5 is NNE, etc.
}
