% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_current_rose.R
\name{adcp_plot_current_rose}
\alias{adcp_plot_current_rose}
\title{Generate current rose}
\usage{
adcp_plot_current_rose(
  dat,
  pal = NULL,
  speed_col = sea_water_speed_cm_s_labels,
  direction_col = sea_water_to_direction_degree_labels,
  speed_label = "Current Speed (cm/s)",
  ncol_legend = 2,
  calculate_prop = TRUE
)
}
\arguments{
\item{dat}{Data frame with at least 2 columns: an ordered factor of direction
groups, and a factor of speed groups. By default, the proportion of
observations in each group (speed and direction) is counted in the
function. If more groups are required, set \code{calculate_prop = FALSE},
and include the proportions in a column called \code{n_prop}. The
proportion is automatically converted to percent for the figure.}

\item{pal}{Vector of colours. Must be the same length as the number of speed
factor levels.}

\item{speed_col}{The column in \code{dat} that holds the speed groups (NOT
QUOTED).}

\item{direction_col}{The column in \code{dat} that holds the direction groups
(NOT QUOTED).}

\item{speed_label}{Title of the current speed legend. Default is "Current
Speed (cm/s)".}

\item{ncol_legend}{Number of columns for the figure legend. Default is 2.}

\item{calculate_prop}{Logical argument. The default, \code{TRUE}, will
calculate the proporation of observations in each speed and direction
group. Set to \code{FALSE} to include proporation in \code{dat}.}
}
\value{
Returns a ggplot object, a rose plot of current speed and direction.
}
\description{
The direction of the petal indicates the direction the current is flowing to.
The colour indicates the current speed. The length of the petal shows the
number of observations for each current speed and direction bin.
}
