% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_current_rose.R
\name{adcp_plot_current_rose}
\alias{adcp_plot_current_rose}
\title{Generate current rose}
\usage{
adcp_plot_current_rose(dat, breaks, speed_cols)
}
\arguments{
\item{dat}{Dataframe that includes columns \code{SPEED} and \code{DIRECTION}.}

\item{breaks}{Number of break points for current speed OR a vector of breaks.
Lower-inclusive.}

\item{speed_cols}{Vector of colours. Must be the same length as
\code{breaks}.}
}
\value{
Returns an "openair" object, a rose plot of current speed and
  direction.
}
\description{
Generate current rose
}
\details{
Generates a current rose using the \code{windRose()} function from
  the \code{openair} package. See help files for \code{openair::windRose} for
  more detail.
}
