% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_current_rose.R
\name{adcp_plot_current_rose}
\alias{adcp_plot_current_rose}
\title{Generate current rose}
\usage{
adcp_plot_current_rose(
  dat,
  pal = NULL,
  speed_col = sea_water_speed_cm_s_labels,
  direction_col = sea_water_to_direction_degree_labels,
  speed_label = "Current Speed (cm/s)"
)
}
\arguments{
\item{dat}{Data frame with at least 3 columns: an ordered factor of direction
groups, a factor of speed groups, and \code{n_prop}, the proportion of
observations in the corresponding direction and speed groups. The
proportion is automatically converted to percent for the figure.}

\item{pal}{Vector of colours. Must be the same length as the number of speed
factor levels.}

\item{speed_col}{The column in \code{dat} that holds the speed groups (NOT
QUOTED).}

\item{direction_col}{The column in \code{dat} that holds the direction groups
(NOT QUOTED).}

\item{speed_label}{Title of the current speed legend. Default is "Current
Speed (cm/s)".}
}
\value{
Returns a ggplot object, a rose plot of current speed and direction.
}
\description{
The direction of the petal indicates the direction the current is flowing to.
The colour indicates the current speed. The length of the petal shows the
number of observations for each current speed and direction bin.
}
