% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_depth_flags.R
\name{adcp_plot_depth}
\alias{adcp_plot_depth}
\title{Plot SENSOR_DEPTH_BELOW_SURFACE}
\usage{
adcp_plot_depth(dat, title = NULL)
}
\arguments{
\item{dat}{Dataframe of ACDP data in long format, as returned by
\code{adcp_format_opendata()}.}

\item{title}{Optional title for the figure.}
}
\value{
ggplot object. Figure shows SENSOR_DEPTH_BELOW_SURFACE over time.
}
\description{
Plot SENSOR_DEPTH_BELOW_SURFACE
}
