% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_speed_at_bins.R
\name{adcp_plot_speed_at_bins}
\alias{adcp_plot_speed_at_bins}
\title{Plot sea water speed through the water column}
\usage{
adcp_plot_speed_at_bins(
  dat,
  summary_col = med_sea_water_speed_cm_s,
  min_col = q_25,
  max_col = q_75,
  shape_col = NULL,
  pal_col = med_sea_water_speed_cm_s_labels,
  pal = NULL,
  pal_label = "Current Speed (cm/s)",
  x_label = "Current Speed (cm/s)"
)
}
\arguments{
\item{dat}{Data frame with columns \code{bin_height_above_sea_floor} (a
factor), \code{summary_col}, \code{min_col}, \code{max_col}, and optionally
\code{shape_col}.}

\item{summary_col}{Column with the summary value (UNQUOTED). Default is
\code{med_sea_water_speed_cm_s}.}

\item{min_col}{Column with the minimum values for the error bar (UNQUOTED).
Default is \code{q_25}.}

\item{max_col}{Column with the maximum values for the error bar (UNQUOTED).
Default is \code{q_75}.}

\item{shape_col}{Column to map onto shape (UNQUOTED).. Optional.}

\item{pal_col}{Column to map to colour (UNQUOTED). MUst be a factor.}

\item{pal}{Option colour palette.}

\item{pal_label}{Character string. Title for the colour legend.}

\item{x_label}{Character string. Title for the x-axis.}
}
\value{
Returns a ggplot object.
}
\description{
Plot sea water speed through the water column
}
