% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_speed_hist.R
\name{adcp_plot_speed_hist}
\alias{adcp_plot_speed_hist}
\title{Plot histogram of speed observations}
\usage{
adcp_plot_speed_hist(dat_hist, bar_cols, speed_label = "Current Speed (cm/s)")
}
\arguments{
\item{dat_hist}{Dataframe that includes columns \code{bins_plot} (text),
\code{prop}, and \code{freq}.}

\item{bar_cols}{Vector of colours. Must be the same length as the number of
bins to plot.}

\item{speed_label}{Title of the current speed legend. Default is "Current
Speed (cm/s)".}
}
\value{
Returns a dataframe of upper and lower bin limits (lower inclusive),
  and frequency and proportion of observations in each bin.
}
\description{
Plot histogram of speed observations
}
