% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_plot_speed_hist.R
\name{adcp_plot_speed_hist}
\alias{adcp_plot_speed_hist}
\title{Plot histogram of speed observations}
\usage{
adcp_plot_speed_hist(
  dat_hist,
  bar_cols,
  text_size = 3,
  speed_label = "Current Speed (cm/s)"
)
}
\arguments{
\item{dat_hist}{Data frame that includes columns \code{bins_plot} (text),
\code{prop}, and \code{freq}.}

\item{bar_cols}{Vector of colours. Must be the same length as the number of
bins to plot.}

\item{text_size}{Size of the text annotating the number of observations in
each bin.}

\item{speed_label}{Title of the current speed legend. Default is "Current
Speed (cm/s)".}
}
\value{
Returns a ggplot object.
}
\description{
Plot histogram of speed observations
}
