% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_read_tracking.R
\name{adcp_read_tracking}
\alias{adcp_read_tracking}
\title{Read deployment details from CMAR tracking sheet}
\usage{
adcp_read_tracking(link = NULL)
}
\arguments{
\item{link}{Link to the Current & Wave Tracking Deployment Details sheet on
Google Drive.}
}
\value{
Returns a data frame of the Deployment Details sheet.
}
\description{
Read deployment details from CMAR tracking sheet
}
