% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_read_txt.R
\name{adcp_read_txt}
\alias{adcp_read_txt}
\title{Read ADCP txt file}
\usage{
adcp_read_txt(path, file_name = NULL)
}
\arguments{
\item{path}{Path to the txt file (including ".txt" extension) or to the
folder where the txt file is saved.}

\item{file_name}{Required if \code{path} does not include the file name.
Include the ".txt" file extension. Default is \code{file_name = NULL}.}
}
\value{
Returns a dataframe of the data with a single header row and each row
  labelled as "SensorDepth", "WaterSpeed", or "WaterDirection".
}
\description{
Read ADCP txt file
}
\details{
The \code{TIMESTAMP} column is in the timezone of the deployment
  date (e.g., "AST" if deployed in November to March and "DST" if deployed in
  March to November). The \code{TIMESTAMP} does NOT account for changes to
  daylight savings time. The \code{TIMESTAMP} is assigned a timezone of "UTC"
  to avoid \code{NA} values during the beginning of daylight savings time
  (e.g., 2019-03-10 02:30:00 is NOT a valid time for the "America/Halifax"
  timezone). This \code{TIMESTAMP} can be converted to true UTC using
  \code{adcp_correct_timestamp()}.
}
