% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp_write_report_table.R
\name{adcp_write_report_table}
\alias{adcp_write_report_table}
\title{Writes deployment table for summary report}
\usage{
adcp_write_report_table(metadata)
}
\arguments{
\item{metadata}{Deployment metadata from the ADCP TRACKING sheet.}
}
\value{
Returns a data frame with columns for the report table.
}
\description{
Writes deployment table for summary report
}
