% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_index}
\alias{find_index}
\title{Index of first bin column}
\usage{
find_index(dat_wide)
}
\arguments{
\item{dat_wide}{Data frame of ADCP data, as exported from
\code{adcp_read_txt()}.}
}
\value{
Returns the index of the first column with bin data.
}
\description{
Index of first bin column
}
